Update base image to debian-base
The ubuntu-slim image has been deprecated and documentation suggests using
debian-base instead.

The new image contains a [`clean-install` script][0] which already takes care
of clearing the cache and running `apt-get clean` etc. So we don't have to do
that here.

`k8s.gcr.io` is an alias for `gcr.io/google_containers`, optimized for
different regions. According [to the PR][1] that introduced the alias, this the
suggested name to use.

[1]: https://github.com/kubernetes/kubernetes/pull/57824