Upgrade Fluentd and all plugins
There's an [issue with the systemd plugin][1] which causes it to corrupt the
position file it uses to track which logs it's seen and which it hasn't which
then causes the Fluentd pod to get stuck in a restart loop. Apparently the
issue has been fixed upstream, so upgrading Fluentd should fix this.

The plugins were updated opportunistically to get any other fixes that may have
been implemented. Didn't see any breaking issues in any of the changelogs.

Changelogs:
* https://github.com/fabric8io/fluent-plugin-kubernetes_metadata_filter/releases
* https://github.com/uken/fluent-plugin-elasticsearch/blob/master/History.md
* https://github.com/repeatedly/fluent-plugin-record-modifier/blob/master/ChangeLog
* https://github.com/reevoo/fluent-plugin-systemd/releases

[1]: https://github.com/reevoo/fluent-plugin-systemd/issues/68